Fix empty user info when matching
